Bug with float type casting SeriesNumber
I've run into the ValueError (thrown by pytables, line 1954) a number of times. This seems to be related to the float type casting, but I have yet to investigate. For the time being, I've simply changed the type for the SeriesNumber column to integer. I suspect this will continue to be an issue with other columns that are casted as float, but will need to investigate further.

The _overwrite_hdf method now resets the data frame index.