Bug fix
Resolved the float serialization issue that would arise when writing the data frame to the HDF5 file. The issue was caused by an additional elif statement that I added to convert empty strings to NaN's for columns being type casted to float. After that replacement, there was no conversion to float.