Check hash logic update
Updated the logic that controls calculation of file hashes on-the-fly. There is no longer a user-specified parameter "checkhash" that controls when/if the hash is verified. Instead, hashes are ignored until all files (not found in the database) are compiled. Then those files previously existing in the database and compiled recently are fed back into "process" to have the hashes checked.